Move CatchOptions to settings, since they are
